// JavaSript/Web/Angular/08. Workshop Forms/workshop/src/app/types/user.ts
export interface User {
  themes: string[];
  posts: string[];
  _id: string;
  tel: string;
  email: string;
  username: string;
  password: string;
  created_at: string;
  updatedAt: string;
  __v: number;
}

export interface UserForAuth {
  id: string;
  firstName: string;
  username: string;
  email: string;
  phoneNumber: string;
  password: string;
}